+++ src/main/java/gift/global/config/RestConfig.java
@@ -1,0 +1,31 @@
+package gift.global.config;
+
+import org.apache.hc.client5.http.impl.classic.HttpClientBuilder;
+import org.apache.hc.client5.http.impl.io.PoolingHttpClientConnectionManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.http.client.ClientHttpRequestFactory;
+import org.springframework.http.client.HttpComponentsClientHttpRequestFactory;
+
+@Configuration
+public class RestConfig {
+
+    @Bean
+    public ClientHttpRequestFactory clientHttpRequestFactory() {
+        var clientHttpRequestFactory = new HttpComponentsClientHttpRequestFactory();
+        clientHttpRequestFactory.setConnectTimeout(3000);
+        clientHttpRequestFactory.setConnectionRequestTimeout(3000);
+
+        var connectionManager = new PoolingHttpClientConnectionManager();
+        connectionManager.setMaxTotal(100);
+        connectionManager.setDefaultMaxPerRoute(10);
+
+        var httpClient = HttpClientBuilder.create()
+            .setConnectionManager(connectionManager)
+            .build();
+
+        clientHttpRequestFactory.setHttpClient(httpClient);
+
+        return clientHttpRequestFactory;
+    }
+}
